App is divided into three modules
